+++ R/ParamUty.R
@@ -6,6 +6,7 @@
 #' @template param_id
 #' @template param_default
 #' @template param_tags
+#' @template param_custom_check
 #'
 #' @family Params
 #' @include Param.R