+++ R/zzz.R
@@ -1,7 +1,7 @@
 #' @import data.table
 #' @import checkmate
 #' @import mlr3misc
-#' @importFrom R6 R6Class
+#' @importFrom R6 R6Class is.R6Class
 #' @importFrom stats runif rnorm
 "_PACKAGE"