+++ NEWS.md
@@ -1,4 +1,9 @@
-# paradox 0.4.0-9000
+# paradox 0.5.0-9000
+
+* `ps()` shortcuts for `ParamSet` construction, with new `Domain` construct and constructors `p_dbl`, `p_int`, `p_lgl`, `p_fct`, and `p_uty`.
+* `ParamSet$tune_ps()` method that constructs tunable `ParamSet` from `TuneToken` objects, which are constructed with `to_tune()`.
+
+# paradox 0.5.0
 
 * Compact in-memory representation of R6 objects to save space when
   saving objects via saveRDS(), serialize() etc.